requirements.txt
keyboard==0.13.5
netifaces==0.10.9
pkg-resources==0.0.0
RPi.GPIO==0.7.0
RPLCD==1.3.0
urllib3==1.25.9
